University_Club_Project_YMH220-main/University_Club_Project_YMH220-main/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
